Refactoring tree structure
